refactor cooking dom updates
